@@ -2,46 +2,43 @@
   "name": "@github/g-emoji-element",
   "version": "1.1.3",
   "description": "Backports native emoji characters to browsers that don't support them by replacing the characters with fallback images.",
-  "main": "dist/index.umd.js",
-  "module": "dist/index.esm.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "type": "module",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "license": "MIT",
   "repository": "github/g-emoji-element",
+  "prettier": "@github/prettier-config",
   "files": [
-    "dist",
-    "index.d.ts"
+    "dist"
   ],
   "scripts": {
     "clean": "rm -rf dist",
     "lint": "github-lint",
-    "prebuild": "npm run clean && npm run lint && mkdir dist",
-    "build": "rollup -c && cp index.js.flow dist/index.esm.js.flow && cp index.js.flow dist/index.umd.js.flow",
+    "prebuild": "npm run clean && npm run lint && mkdir dist && tsc",
+    "build": "rollup -c",
     "pretest": "npm run build",
-    "generate-emoji": "script/generate-emoji-modifiers > modifiers.js && prettier --write modifiers.js",
-    "test": "karma start test/karma.config.js",
+    "generate-emoji": "script/generate-emoji-modifiers.cjs > src/modifiers.ts && prettier --write src/modifiers.ts",
+    "test": "karma start test/karma.config.cjs",
     "prepublishOnly": "npm run build",
     "postpublish": "npm publish --ignore-scripts --@github:registry='https://npm.pkg.github.com'"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.5",
-    "babel-preset-github": "^3.2.0",
+    "@github/prettier-config": "0.0.4",
     "chai": "^4.2.0",
-    "eslint": "^6.6.0",
+    "eslint": "^6.8.0",
     "eslint-plugin-github": "^2.0.0",
-    "flow-bin": "^0.101.1",
-    "karma": "^4.1.0",
+    "karma": "^5.0.9",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-mocha": "^1.3.0",
+    "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "mocha": "^6.1.4",
-    "rollup": "^1.26.3",
-    "rollup-plugin-babel": "^4.3.3",
+    "mocha": "^6.2.3",
+    "rollup": "^1.32.1",
+    "typescript": "^3.9.3",
     "unicode-emoji-json": "0.0.2"
   },
   "eslintIgnore": [
     "dist/"
-  ],
-  "dependencies": {}
+  ]
 }
