@@ -1,5 +1,3 @@
-/* @flow strict */
-
 import {isEmojiSupported} from './emoji-detection'
 import {applyTone, applyTones, removeTone} from './tones'
 
@@ -41,7 +39,7 @@ class GEmojiElement extends HTMLElement {
     }
   }
 
-  static get observedAttributes(): Array<string> {
+  static get observedAttributes(): string[] {
     return ['tone']
   }
 
@@ -59,12 +57,12 @@ function updateTone(el: GEmojiElement) {
 
   const tones = el.tone.split(' ').map(x => parseInt(x, 10))
   if (tones.length === 0) {
-    el.textContent = removeTone(el.textContent)
+    el.textContent = removeTone(el.textContent || '')
   } else if (tones.length === 1) {
     const tone = tones[0]
-    el.textContent = tone === 0 ? removeTone(el.textContent) : applyTone(el.textContent, tone)
+    el.textContent = tone === 0 ? removeTone(el.textContent || '') : applyTone(el.textContent || '', tone)
   } else {
-    el.textContent = applyTones(el.textContent, tones)
+    el.textContent = applyTones(el.textContent || '', tones)
   }
 }
 
@@ -73,7 +71,7 @@ function updateTone(el: GEmojiElement) {
 // el - The <g-emoji> element.
 //
 // Returns an HTMLImageElement.
-function emojiImage(el) {
+function emojiImage(el: Element) {
   const image = document.createElement('img')
   image.className = 'emoji'
   image.alt = el.getAttribute('alias') || ''
@@ -83,6 +81,12 @@ function emojiImage(el) {
 }
 
 export default GEmojiElement
+
+declare global {
+  interface Window {
+    GEmojiElement: typeof GEmojiElement
+  }
+}
 
 if (!window.customElements.get('g-emoji')) {
   window.GEmojiElement = GEmojiElement
