@@ -1,7 +1,7 @@
 module.exports = function(config) {
   config.set({
     frameworks: ['mocha', 'chai'],
-    files: ['../dist/index.umd.js', 'test.js'],
+    files: [{pattern: '../dist/index.js', type: 'module'}, 'test.js'],
     reporters: ['mocha'],
     port: 9876,
     colors: true,
